requirements.txt
openai>=1.30.0
google-genai>=1.0.0
pydantic>=2.5.0
pdf2image>=1.16.3
Pillow>=10.0.0
python-dotenv>=1.0.0
rich>=13.7.0
click>=8.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
sqlalchemy>=2.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.0
bcrypt>=4.0.0
gunicorn>=21.2.0